feat: move v2 to main page
